...fix ketchup cancel for linux
NOTE: In Linux, once a `tomato_job` completes or get cancelled, its
children will be closed/killed, and it will remain a zombie process.
There is no clear way to kill it directly, but everything seems to work
anyway, it does not do any harm.
I noticed that this zombie process terminates once another tomato_job
starts or the daemon is terminated.